implement buffer for receiving strings over UART
Our communication protocol will have 5 characters, e.g., t0435, where
the first character (t or s) indicates which PMT we're controlling, then
the other 4 characters indicate the voltage in mV. The UART buffer can
hold 7 characters, because the sender is expected to send a carriage
return (\r) at the end of the string, then we add a null character to
the end to null-terminate the string.

This commit adds the functionality for buffering the received message
(up to 6 characters) and ensuring that buffer overflow doesn't occur.

The UART rx buffer consumers will check if the buffer is full
(rxbuf_is_full()), then call a getter to receive a copy of the buffer.
If they call the getter when the buffer isn't full, we return an error
code.