add pmt configuration header
This header file contains all the configuration needed to specify the
PMTs in the system. The parser will use this information to validate the
voltage-set commands.

This configuration should be generic enough that any number of PMTs can
be supported. The particular MSP430 we're using only has 4 DACs, so that
places a limit of 4 PMTs in practice.